requirements.txt
openai==0.27.9
numpy==1.23.5
pandas==1.3.2
requests==2.31.0
bs4==0.0.1
tqdm
pytest==7.4.0
yfinance==0.2.28
seaborn==0.12.2
hydra-core==1.3.2
omegaconf==2.3.0
tiktoken==0.5.1
python-dotenv==1.0.0
imageio==2.31.3
scikit-image==0.21.0
